backendM/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.11
dj-database-url==3.1.2
whitenoise==6.12.0
gunicorn==25.3.0
Pillow==11.0.0
